refactor: mv ValidateRayServiceSpec func and unit test to validation.go
