+++ ray-operator/controllers/ray/rayservice_controller_unit_test.go
@@ -26,38 +26,6 @@ import (
 	"github.com/ray-project/kuberay/ray-operator/test/support"
 )
 
-func TestValidateRayServiceSpec(t *testing.T) {
-	err := validateRayServiceSpec(&rayv1.RayService{
-		Spec: rayv1.RayServiceSpec{
-			RayClusterSpec: rayv1.RayClusterSpec{
-				HeadGroupSpec: rayv1.HeadGroupSpec{
-					HeadService: &corev1.Service{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "my-head-service",
-						},
-					},
-				},
-			},
-		},
-	})
-	assert.Error(t, err, "spec.rayClusterConfig.headGroupSpec.headService.metadata.name should not be set")
-
-	err = validateRayServiceSpec(&rayv1.RayService{
-		Spec: rayv1.RayServiceSpec{},
-	})
-	assert.NoError(t, err, "The RayService spec is valid.")
-
-	var upgradeStrat rayv1.RayServiceUpgradeType = "invalidStrategy"
-	err = validateRayServiceSpec(&rayv1.RayService{
-		Spec: rayv1.RayServiceSpec{
-			UpgradeStrategy: &rayv1.RayServiceUpgradeStrategy{
-				Type: &upgradeStrat,
-			},
-		},
-	})
-	assert.Error(t, err, "spec.UpgradeSpec.Type is invalid")
-}
-
 func TestGenerateHashWithoutReplicasAndWorkersToDelete(t *testing.T) {
 	// `generateRayClusterJsonHash` will mute fields that will not trigger new RayCluster preparation. For example,
 	// Autoscaler will update `Replicas` and `WorkersToDelete` when scaling up/down. Hence, `hash1` should be equal to
